builds address cleaner PORO, deletes code in prepare controller
